+++ src/__stories__/__data__/other/index.ts
@@ -4,3 +4,4 @@ export * from './legend-position';
 export * from './line-and-bar';
 export * from './lines';
 export * from './tooltip';
+export * from './legend';